+++ InternetArchive.NET/Wayback.cs
@@ -2,8 +2,10 @@
 
 public class Wayback(Client client)
 {
-    private const string WaybackUrl = "https://archive.org/wayback/available";
     private const string CdxUrl = "https://web.archive.org/cdx/search/cdx";
+    private const string SavePageUrl = "https://web.archive.org/save";
+    private static string SavePageGetJobStatusUrl(string jobId) => $"https://web.archive.org/save/status/{jobId}";
+    private static string SavePageGetSystemStatusUrl => $"https://web.archive.org/save/status/system";
 
     internal static readonly string DateFormat = "yyyyMMddHHmmss";
 
@@ -17,7 +19,7 @@ public class Wayback(Client client)
         public string? MatchType { get; set; }
         public string? Collapse { get; set; }
         public int? Limit { get; set; }
-        [Obsolete] public int? Offset { get; set; }
+        [Obsolete("Support removed from archive.org in 2024")] public int? Offset { get; set; }
         public int? Page { get; set; }
         public int? PageSize { get; set; }
         public bool FastLatest { get; set; }
@@ -26,9 +28,9 @@ public class Wayback(Client client)
         internal Dictionary<string, string> ToQuery()
         {
             if (Url == null) throw new InternetArchiveException("Url is required");
-#pragma warning disable CS0612 // Type or member is obsolete
+#pragma warning disable CS0618 // Type or member is obsolete
             if (Offset.HasValue) throw new InternetArchiveException("Offset is no longer supported");
-#pragma warning restore CS0612
+#pragma warning restore CS0618
 
             var query = new Dictionary<string, string> { { "url", Url }, { "showResumeKey", "true" } };
 
@@ -103,4 +105,153 @@ public class Wayback(Client client)
         return response;
     }
 
+    public class SavePageRequest
+    {
+        [JsonPropertyName("url")]
+        public string? Url { get; set; }
+
+        [JsonPropertyName("capture_all")]
+        public bool? CaptureAll { get; set; }
+
+        [JsonPropertyName("capture_outlinks")]
+        public bool CaptureOutlinks { get; set; } = false;
+
+        [JsonPropertyName("capture_screenshot")]
+        public bool? CaptureScreenshot { get; set; }
+
+        [JsonPropertyName("delay_wb_availability")]
+        public bool? DelayAvailability { get; set; }
+
+        [JsonPropertyName("force_get")]
+        public bool? ForceGet { get; set; }
+
+        [JsonPropertyName("skip_first_archive")]
+        public bool SkipFirstArchive { get; set; } = true;
+
+        [JsonPropertyName("if_not_archived_within")]
+        public string? IfNotArchivedWithin { get; set; }
+
+        [JsonPropertyName("outlinks_availability")]
+        public bool? OutlinksAvailability { get; set; }
+
+        [JsonPropertyName("email_result")]
+        public bool? EmailResult { get; set; }
+
+        [JsonPropertyName("js_behavior_timeout")]
+        public int? JavascriptTimeout { get; set; }
+
+        [JsonPropertyName("capture_cookie")]
+        public string? CaptureCookie { get; set; }
+
+        [JsonPropertyName("use_user_agent")]
+        public string? UseUserAgent { get; set; }
+
+        [JsonPropertyName("target_username")]
+        public string? TargetUsername { get; set; }
+
+        [JsonPropertyName("target_password")]
+        public string? TargetPassword { get; set; }
+    }
+
+    public class SavePageResponse
+    {
+        [JsonPropertyName("url")]
+        public string? Url { get; set; }
+
+        [JsonPropertyName("job_id")]
+        public string? JobId { get; set; }
+
+        [JsonPropertyName("message")]
+        public string? Message { get; set; }
+    }
+
+    public async Task<SavePageResponse?> SavePageAsync(SavePageRequest request, CancellationToken cancellationToken = default)
+    {
+        if (request.Url == null) throw new InternetArchiveException("Url is required");
+
+        var form = new Dictionary<string, object>();
+        foreach (var property in request.GetType().GetProperties())
+        {
+            var attribute = property.GetCustomAttribute<JsonPropertyNameAttribute>();
+            var key = attribute?.Name ?? throw new Exception("JsonPropertyNameAttribute is required");
+            var value = property.GetValue(request);
+            if (value == null) continue;
+
+            if (property.PropertyType == typeof(bool) || property.PropertyType == typeof(bool?))
+            {
+                if ((bool)value == false) value = 0;
+                else value = 1;
+            }
+
+            form.Add(key, value);
+        }
+
+        var requestHeaders = new Dictionary<string, string?> { { "Accept", "application/json" } };
+        var response = await _client.SendAsync<SavePageResponse>(HttpMethod.Post, SavePageUrl, form, requestHeaders, cancellationToken).ConfigureAwait(false);
+
+        return response;
+    }
+
+    public class SavePageStatusResponse
+    {
+        [JsonPropertyName("status")]
+        public string? Status { get; set; }
+
+        [JsonPropertyName("exception")]
+        public string? Exception { get; set; }
+
+        [JsonPropertyName("status_ext")]
+        public string? StatusException { get; set; }
+
+        [JsonPropertyName("job_id")]
+        public string? JobId { get; set; }
+
+        [JsonPropertyName("original_url")]
+        public string? OriginalUrl { get; set; }
+
+        [JsonPropertyName("screenshot")]
+        public string? Screenshot { get; set; }
+
+        [JsonPropertyName("timestamp")]
+        public string? Timestamp { get; set; }
+
+        [JsonPropertyName("duration_sec")]
+        public double? DurationSeconds { get; set; }
+
+        [JsonPropertyName("resources")]
+        public IEnumerable<string>? Resources { get; set; }
+
+        public class Outlink
+        {
+            [JsonPropertyName("url")]
+            public string? Url { get; set; }
+
+            [JsonPropertyName("job_id")]
+            public string? JobId { get; set; }
+        }
+
+        [JsonPropertyName("outlinks")]
+        public IEnumerable<Outlink>? Outlinks { get; set; }
+    }
+
+    public async Task<SavePageStatusResponse?> GetSavePageStatusAsync(string jobId, CancellationToken cancellationToken = default)
+    {
+        var response = await _client.GetAsync<SavePageStatusResponse>(SavePageGetJobStatusUrl(jobId), cancellationToken).ConfigureAwait(false);
+        if (response?.Status == "error") throw new InternetArchiveResponseException(response.StatusException ?? "error");
+        return response;
+    }
+
+    public class SavePageSystemStatusResponse
+    {
+        [JsonPropertyName("status")]
+        public string? Status { get; set; }
+
+        [JsonPropertyName("recent_captures")]
+        public int? RecentCaptures { get; set; }
+    }
+
+    public async Task<SavePageSystemStatusResponse?> GetSavePageSystemStatusAsync(CancellationToken cancellationToken = default)
+    {
+        return await _client.GetAsync<SavePageSystemStatusResponse>(SavePageGetSystemStatusUrl, cancellationToken).ConfigureAwait(false);
+    }
 }